Seperate out S3Error.
