+++ core/server/proxy/src/main/java/alluxio/proxy/s3/S3RestUtils.java
@@ -21,8 +21,6 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 
 import javax.ws.rs.core.Response;
-import javax.xml.bind.annotation.XmlElement;
-import javax.xml.bind.annotation.XmlRootElement;
 
 /**
  * Utilities for handling S3 REST calls.
@@ -91,77 +89,8 @@ public final class S3RestUtils {
    * @return the response
    */
   private static Response createErrorResponse(S3Exception e) {
-    Error errorResponse = new Error(e.getResource(), e.getErrorCode());
+    S3Error errorResponse = new S3Error(e.getResource(), e.getErrorCode());
     return Response.status(e.getErrorCode().getStatus()).entity(errorResponse).build();
-  }
-
-  /**
-   * Error response defined in http://docs.aws.amazon.com/AmazonS3/latest/API/ErrorResponses.html.
-   * It will be encoded into an XML string to be returned as an error response for the REST call.
-   */
-  @XmlRootElement(name = "Error")
-  public static class Error {
-    private final String mCode;
-    private final String mMessage;
-    private final String mRequestId;
-    private final String mResource;
-
-    /**
-     * Creates an {@link Error}.
-     * This is needed for {@link Error} to be encoded into XML, see <a>https://stackoverflow.com/
-     * questions/26014184/jersey-returns-500-when-trying-to-return-an-xml-response</a> for details.
-     */
-    public Error() {
-      mCode = "";
-      mMessage = "";
-      mRequestId = "";
-      mResource = "";
-    }
-
-    /**
-     * Creates a new {@link Error}.
-     *
-     * @param resource the resource (bucket or object key) where the error happens
-     * @param code the error code
-     */
-    public Error(String resource, S3ErrorCode code) {
-      mCode = code.getCode();
-      mMessage = code.getDescription();
-      mRequestId = "";
-      mResource = resource;
-    }
-
-    /**
-     * @return the error code
-     */
-    @XmlElement(name = "Code")
-    public String getCode() {
-      return mCode;
-    }
-
-    /**
-     * @return the error message
-     */
-    @XmlElement(name = "Message")
-    public String getMessage() {
-      return mMessage;
-    }
-
-    /**
-     * @return the request ID
-     */
-    @XmlElement(name = "RequestId")
-    public String getRequestId() {
-      return mRequestId;
-    }
-
-    /**
-     * @return the resource name
-     */
-    @XmlElement(name = "Resource")
-    public String getResource() {
-      return mResource;
-    }
   }
 
   private S3RestUtils() {} // prevent instantiation
